Start moving to strategy object (more conservative)
